requirements.txt
# Core ML & Computer Vision
numpy==1.24.3
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
Pillow==10.1.0

# Deep Learning
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.200

# Model Conversion & Optimization
onnx==1.15.0
onnxruntime==1.16.3
tensorflow==2.14.0

# Data Augmentation & Processing
albumentations==1.3.1
imgaug==0.4.0
scikit-image==0.22.0

# Dataset Management
pandas==2.1.3
tqdm==4.66.1
xmltodict==0.13.0

# Visualization & Analysis
matplotlib==3.8.2
seaborn==0.13.0

# Camera & Device Access
pygame==2.5.2

# Utility
pyyaml==6.0.1
scipy==1.11.4
